+++ cpp/src/layout/barnes_hut.hpp
@@ -17,16 +17,15 @@
 #pragma once
 
 #include <rmm/thrust_rmm_allocator.h>
-#include <rmm_utils.h>
+#include <utilities/error_utils.h>
+
 #include <stdio.h>
 #include <converters/COOtoCSR.cuh>
 #include <graph.hpp>
 #include <internals.hpp>
-#include <rmm/device_buffer.hpp>
 
 #include "bh_kernels.hpp"
 #include "fa2_kernels.hpp"
-#include "utilities/error_utils.h"
 #include "utilities/graph_utils.cuh"
 #include "utils.hpp"