+++ cpp/tests/centrality/betweenness_centrality_test.cu
@@ -29,6 +29,7 @@
 
 #include <fstream>
 
+#include <rmm/mr/device/cuda_memory_resource.hpp>
 #include "traversal/bfs_ref.h"
 
 #ifndef TEST_EPSILON
@@ -367,9 +368,9 @@ INSTANTIATE_TEST_CASE_P(simple_test,
 
 int main(int argc, char **argv)
 {
-  rmmInitialize(nullptr);
   testing::InitGoogleTest(&argc, argv);
+  auto resource = std::make_unique<rmm::mr::cuda_memory_resource>();
+  rmm::mr::set_default_resource(resource.get());
   int rc = RUN_ALL_TESTS();
-  rmmFinalize();
   return rc;
 }