+++ python/cugraph/link_prediction/overlap_wrapper.pyx
@@ -21,7 +21,6 @@ from cugraph.link_prediction.overlap cimport overlap_list as c_overlap_list
 from cugraph.structure.graph_new cimport *
 from cugraph.structure import graph_new_wrapper
 from libc.stdint cimport uintptr_t
-from libc.stdlib cimport calloc, malloc, free
 from cython cimport floating
 
 import cudf